requirements.txt
Django>=5.1,<5.2
django-allauth>=65.0,<66.0
django-htmx>=1.19,<2.0
requests>=2.32,<3.0
gunicorn>=23.0,<24.0
python-dotenv>=1.0,<2.0
whitenoise>=6.7,<7.0
